refactor: bot framework registration (#8992)
* refactor: resource bot service

* refactor: test

* refactor: rename

* refactor: rename

* test: add test
